+++ subprojects/testfx-core/src/main/java/org/testfx/matcher/base/ParentMatchers.java
@@ -8,7 +8,7 @@ import org.hamcrest.Matcher;
 public class ParentMatchers {
 
     //---------------------------------------------------------------------------------------------
-    // MATCHER METHODS.
+    // STATIC FACTORY METHODS.
     //---------------------------------------------------------------------------------------------
 
     @Factory
